1. Use Data Binding to change Layout 2. Adjust the Naming style and part of useless datablocks
