+++ android/canonical/app/src/main/java/com/google/samples/quickstart/canonical/MainActivity.kt
@@ -1,32 +1,9 @@
 package com.google.samples.quickstart.canonical
 
-import android.content.pm.PackageManager
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.FragmentTransaction
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class StopwatchView : ViewModel(){
-
-  var pauseOffset = MutableLiveData<Long>(0L)
-  var isWorking = MutableLiveData<Boolean>(false)
-  var fragmentPauseStartTime = MutableLiveData<Long>(0L)
-
-  fun setPauseOffset(pause_offset_value:Long){
-    pauseOffset.value = pause_offset_value
-  }
-
-  fun setWorkingStatus(working_status:Boolean){
-    isWorking.value = working_status
-  }
-
-  fun setFragmentPauseStartTime(fragment_pause_start_time:Long){
-    fragmentPauseStartTime.value = fragment_pause_start_time
-  }
-}
 
 class MainActivity : AppCompatActivity() {
 
@@ -42,7 +19,7 @@ class MainActivity : AppCompatActivity() {
     runFragment = RunFragment()
     supportFragmentManager
       .beginTransaction()
-      .replace(R.id.fragment_layout, runFragment)
+      .replace(R.id.fragment_container, runFragment)
       .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_OPEN)
       .commit()
 
@@ -54,33 +31,34 @@ class MainActivity : AppCompatActivity() {
           runFragment = RunFragment()
           supportFragmentManager
             .beginTransaction()
-            .replace(R.id.fragment_layout, runFragment)
+            .replace(R.id.fragment_container, runFragment)
             .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_OPEN)
             .commit()
+          true
         }
 
         R.id.bottom_navigation_item_map -> {
           mapsFragment = MapsFragment()
           supportFragmentManager
             .beginTransaction()
-            .replace(R.id.fragment_layout, mapsFragment)
+            .replace(R.id.fragment_container, mapsFragment)
             .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_OPEN)
             .commit()
+          true
         }
 
         R.id.bottom_navigation_item_profile -> {
           meFragment = MeFragment()
           supportFragmentManager
             .beginTransaction()
-            .replace(R.id.fragment_layout, meFragment)
+            .replace(R.id.fragment_container, meFragment)
             .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_OPEN)
             .commit()
+          true
         }
 
-
+        else -> false
       }
-      true
-
     }
   }
