+++ android/canonical/app/src/main/res/menu/bottom_navigation_menu.xml
@@ -3,16 +3,16 @@
     <item
         android:id="@+id/bottom_navigation_item_run"
         android:icon="@drawable/ic_baseline_directions_run_24"
-        android:title="@string/navigationMenuRun" />
+        android:title="@string/navigation_menu_run" />
 
     <item
         android:id="@+id/bottom_navigation_item_map"
         android:icon="@drawable/ic_baseline_map_24"
-        android:title="@string/navigationMenuMap" />
+        android:title="@string/navigation_menu_map" />
 
     <item
         android:id="@+id/bottom_navigation_item_profile"
         android:icon="@drawable/ic_baseline_people_24"
-        android:title="@string/navigationMenuMe"/>
+        android:title="@string/navigation_menu_me"/>
 
 </menu>
