+++ android/canonical/app/src/main/res/values/strings.xml
@@ -1,9 +1,8 @@
 <resources>
     <string name="app_name">Summer 2020</string>
-    <string name="navigationMenuRun">Run</string>
-    <string name="navigationMenuMap">Map</string>
-    <string name="navigationMenuMe">Me</string>
-    <!-- TODO: Remove or change this placeholder text -->
+    <string name="navigation_menu_run">Run</string>
+    <string name="navigation_menu_map">Map</string>
+    <string name="navigation_menu_me">Me</string>
     <string name="hello_blank_fragment">This is Run fragment</string>
     <string name="stop">Stop Timer</string>
     <string name="start">Start Timer</string>
